getLifetime refactoring; new shader from shadertoy :)

+++ include/SoloDevice.h
@@ -44,8 +44,8 @@ namespace solo
 
 		virtual Vector2 getCanvasSize() const = 0;
 
-		/// Returns time elapsed since the engine startup in ms
-		virtual unsigned long getLifetime() const = 0;
+		/// Returns time elapsed since the engine startup in seconds
+		virtual float getLifetime() const = 0;
 
 		virtual void beginUpdate() = 0;
 		virtual void endUpdate() = 0;
